@@ -292,7 +292,7 @@ class GridDefinition(dict):
             )
         kwargs = {}
         if self["cell_width"] is not None:
-            kwargs["resolution"] = (self["cell_width"], self["cell_height"])
+            kwargs["resolution"] = (abs(self["cell_width"]), abs(self["cell_height"]))
         return DynamicAreaDefinition(
             self["grid_name"], self["grid_name"], self.proj4_dict, self["width"], self["height"], **kwargs
         )
